+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/beregningsgrunnlag/soeskenjustering/SoeskenjusteringPeriode.tsx
@@ -3,7 +3,7 @@ import { behandlingErRedigerbar } from '~components/behandling/felles/utils'
 import PeriodeAccordion from '~components/behandling/beregningsgrunnlag/PeriodeAccordion'
 import MaanedVelger from '~components/behandling/beregningsgrunnlag/MaanedVelger'
 import { BodyShort, Button, HStack, Label, Radio, RadioGroup } from '@navikt/ds-react'
-import { Soesken } from '~components/behandling/soeknadsoversikt/familieforhold/personer/Soesken'
+import { Soesken } from '~components/behandling/beregningsgrunnlag/soeskenjustering/Soesken'
 import React from 'react'
 import { IBehandlingReducer } from '~store/reducers/BehandlingReducer'
 import { Familieforhold, IPdlPerson } from '~shared/types/Person'
