+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/Soeknadoversikt.tsx
@@ -1,4 +1,3 @@
-import { Familieforhold } from './familieforhold/Familieforhold'
 import { Box, Heading, HStack } from '@navikt/ds-react'
 import { BehandlingHandlingKnapper } from '../handlinger/BehandlingHandlingKnapper'
 import { NesteOgTilbake } from '../handlinger/NesteOgTilbake'
@@ -25,7 +24,7 @@ import { ViderefoereOpphoer } from '~components/behandling/soeknadsoversikt/vide
 import { TidligereFamiliepleier } from '~components/behandling/soeknadsoversikt/tidligereFamiliepleier/TidligereFamiliepleier'
 import SluttBehandlingOmgjoering from '~components/behandling/soeknadsoversikt/SluttbehandlingOmgjoering'
 import { SoeknadInformasjon } from '~components/behandling/soeknadsoversikt/SoeknadInformasjon'
-import { ForbedretFamilieforhold } from '~components/behandling/soeknadsoversikt/familieforhold/ForbedretFamilieforhold'
+import { Familieforhold } from '~components/behandling/soeknadsoversikt/familieforhold/Familieforhold'
 
 export const Soeknadsoversikt = ({ behandling }: { behandling: IDetaljertBehandling }) => {
   const innloggetSaksbehandler = useInnloggetSaksbehandler()
@@ -52,14 +51,6 @@ export const Soeknadsoversikt = ({ behandling }: { behandling: IDetaljertBehandl
       </Box>
 
       <Box paddingBlock="4" borderWidth="0 0 1 0" borderColor="border-subtle">
-        <ForbedretFamilieforhold
-          behandling={behandling}
-          personopplysninger={personopplysninger}
-          redigerbar={redigerbar}
-        />
-      </Box>
-
-      <Box paddingBlock="4 0" borderWidth="0 0 1 0" borderColor="border-subtle">
         <Familieforhold behandling={behandling} personopplysninger={personopplysninger} redigerbar={redigerbar} />
       </Box>
