+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/familieforhold/Familieforhold.tsx
@@ -1,69 +1,105 @@
-import { IDetaljertBehandling } from '~shared/types/IDetaljertBehandling'
-import { SakType } from '~shared/types/sak'
-import { FamilieforholdBarnepensjon } from '~components/behandling/soeknadsoversikt/familieforhold/barnepensjon/FamilieforholdBarnepensjon'
-import { FamilieforholdOmstillingsstoenad } from '~components/behandling/soeknadsoversikt/familieforhold/omstillingsstoenad/FamilieforholdOmstillingsstoenad'
-import { Box, Heading } from '@navikt/ds-react'
-import React, { useEffect } from 'react'
-import { Personopplysninger } from '~shared/types/grunnlag'
-import styled from 'styled-components'
-import { RedigerFamilieforholdModal } from '~components/behandling/soeknadsoversikt/familieforhold/RedigerFamilieforholdModal'
 import { useApiCall } from '~shared/hooks/useApiCall'
-import { mapApiResult, Result } from '~shared/api/apiUtils'
-import Spinner from '~shared/Spinner'
-import { finnLandSomTekst } from '~components/person/personopplysninger/utils'
 import { hentAlleLand } from '~shared/api/behandling'
-import { ILand } from '~utils/kodeverk'
+import { useEffect } from 'react'
+import { Box, Heading, HStack, VStack } from '@navikt/ds-react'
+import { IDetaljertBehandling } from '~shared/types/IDetaljertBehandling'
+import { Personopplysninger } from '~shared/types/grunnlag'
+import { PersonButtonLink } from '~components/person/lenker/PersonButtonLink'
+import { PersonOversiktFane } from '~components/person/Person'
+import { ExternalLinkIcon } from '@navikt/aksel-icons'
+import { RedigerFamilieforholdModal } from '~components/behandling/soeknadsoversikt/familieforhold/RedigerFamilieforholdModal'
+import { mapResult } from '~shared/api/apiUtils'
+import Spinner from '~shared/Spinner'
+import { ApiErrorAlert } from '~ErrorBoundary'
+import { TabellOverAvdoede } from '~components/behandling/soeknadsoversikt/familieforhold/TabellOverAvdoede'
+import { TabellOverGjenlevende } from '~components/behandling/soeknadsoversikt/familieforhold/TabellOverGjenlevende'
+import { SakType } from '~shared/types/sak'
+import { TabellOverAvdoedesBarn } from '~components/behandling/soeknadsoversikt/familieforhold/TabellOverAvdoedesBarn'
+import { hentLevendeSoeskenFraAvdoedeForSoekerGrunnlag } from '~shared/types/Person'
+import { AnnenForelderSkjema } from '~components/behandling/soeknadsoversikt/familieforhold/AnnenForelderSkjema'
+import { SamsvarPersongalleri } from '~components/behandling/soeknadsoversikt/familieforhold/SamsvarPersongalleri'
 
-export interface PropsFamilieforhold {
+interface Props {
   behandling: IDetaljertBehandling
-  personopplysninger: Personopplysninger | null
   redigerbar: boolean
+  personopplysninger: Personopplysninger | null
 }
 
-export const Familieforhold = ({ behandling, personopplysninger, redigerbar }: PropsFamilieforhold) => {
-  const [landListeResult, landListeFetch] = useApiCall(hentAlleLand)
+export const Familieforhold = ({ behandling, redigerbar, personopplysninger }: Props) => {
+  const [hentAlleLandResult, hentAlleLandFetch] = useApiCall(hentAlleLand)
+
+  const skaViseAnnenForelderSkjema =
+    behandling.sakType === SakType.BARNEPENSJON &&
+    personopplysninger?.avdoede.length === 1 &&
+    personopplysninger.gjenlevende.length === 0
+
   useEffect(() => {
-    landListeFetch(null)
+    hentAlleLandFetch(undefined)
   }, [])
 
   return (
-    <>
-      <Box paddingInline="16" paddingBlock="4">
-        <Heading spacing size="medium" level="2" as="div">
+    <VStack paddingInline="16" paddingBlock="4" gap="8">
+      <HStack gap="4" justify="start" align="center" wrap={false}>
+        <Heading size="medium" level="2">
           Familieforhold
-          {personopplysninger && redigerbar && (
-            <RedigerFamilieforholdModal behandling={behandling} personopplysninger={personopplysninger} />
-          )}
         </Heading>
-      </Box>
+        <div>
+          <PersonButtonLink
+            variant="secondary"
+            size="small"
+            icon={<ExternalLinkIcon aria-hidden />}
+            fnr={personopplysninger?.soeker?.opplysning.foedselsnummer ?? ''}
+            fane={PersonOversiktFane.PERSONOPPLYSNINGER}
+            target="_blank"
+          >
+            Åpne personopplysninger i eget vindu
+          </PersonButtonLink>
+        </div>
 
-      <FamilieforholdWrapper>
-        {behandling.sakType === SakType.BARNEPENSJON ? (
-          <FamilieforholdBarnepensjon
-            personopplysninger={personopplysninger}
-            landListeResult={landListeResult}
-            behandlingId={behandling.id}
-          />
-        ) : (
-          <FamilieforholdOmstillingsstoenad personopplysninger={personopplysninger} landListeResult={landListeResult} />
+        {redigerbar && personopplysninger && (
+          <RedigerFamilieforholdModal behandling={behandling} personopplysninger={personopplysninger} />
         )}
-      </FamilieforholdWrapper>
-    </>
+      </HStack>
+      <Box paddingInline="4 0">
+        {mapResult(hentAlleLandResult, {
+          pending: <Spinner label="Henter alle land..." />,
+          error: (error) => <ApiErrorAlert>{error.detail || 'Kunne ikke hente alle land'}</ApiErrorAlert>,
+          success: (alleLand) => (
+            <VStack gap="8">
+              <SamsvarPersongalleri landListe={alleLand} />
+              {skaViseAnnenForelderSkjema && (
+                <AnnenForelderSkjema behandlingId={behandling.id} personopplysninger={personopplysninger} />
+              )}
+              <TabellOverAvdoede avdoede={personopplysninger?.avdoede} alleLand={alleLand} />
+              {behandling.sakType === SakType.OMSTILLINGSSTOENAD ? (
+                <>
+                  <TabellOverGjenlevende
+                    gjenlevende={!!personopplysninger?.soeker ? [personopplysninger?.soeker] : []}
+                    alleLand={alleLand}
+                  />
+                  <TabellOverAvdoedesBarn
+                    avdoedesBarn={personopplysninger?.avdoede.flatMap((it) => it.opplysning.avdoedesBarn ?? [])}
+                    soeker={personopplysninger?.soeker}
+                    sakType={behandling.sakType}
+                  />
+                </>
+              ) : (
+                <>
+                  <TabellOverGjenlevende gjenlevende={personopplysninger?.gjenlevende} alleLand={alleLand} />
+                  <TabellOverAvdoedesBarn
+                    avdoedesBarn={hentLevendeSoeskenFraAvdoedeForSoekerGrunnlag(
+                      personopplysninger?.avdoede ?? [],
+                      personopplysninger?.soeker?.opplysning.foedselsnummer ?? ''
+                    )}
+                    soeker={personopplysninger?.soeker}
+                    sakType={behandling.sakType}
+                  />
+                </>
+              )}
+            </VStack>
+          ),
+        })}
+      </Box>
+    </VStack>
   )
 }
-
-export const visLandInfoFraKodeverkEllerDefault = (landListeResult: Result<ILand[]>, statsborgerskap?: string) => {
-  return mapApiResult(
-    landListeResult,
-    <Spinner label="Henter landliste" />,
-    () => <>{statsborgerskap ?? 'Ukjent'}</>,
-    (landListe) => <>{statsborgerskap ? finnLandSomTekst(statsborgerskap, landListe) : 'Ukjent'}</>
-  )
-}
-
-export const FamilieforholdWrapper = styled.div`
-  padding: 1em 4em;
-  display: grid;
-  gap: 4rem;
-  margin-bottom: 4rem;
-`
