--- apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/familieforhold/personer/personinfo/PersonInfoFnr.tsx
@@ -1,16 +1,0 @@
-import { Box } from '@navikt/ds-react'
-
-type Props = {
-  fnr: string | undefined
-}
-
-export const PersonInfoFnr = ({ fnr }: Props) => {
-  return (
-    <Box paddingBlock="2 0">
-      <div>
-        <strong>Fødselsnummer</strong>
-      </div>
-      {fnr}
-    </Box>
-  )
-}
